+++ app/src/main/res/layout/information_movie.xml
@@ -74,11 +74,14 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@id/second_line" />
 
-    <com.example.searchmovie.presentation.customView.DescriptionReadMore
+    <TextView
         android:id="@+id/description_movie"
-        android:layout_width="wrap_content"
+        android:layout_width="match_parent"
         android:layout_height="wrap_content"
+        android:layout_margin="15dp"
+        android:text=""
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@id/description" />
+
 </androidx.constraintlayout.widget.ConstraintLayout>
